pyproject.toml
[project]
name = "timetracker"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Movsisyan", email = "movsisyan@protonmail.com" }
]
dependencies = [
    "pyscreeze>=0.1.30",
    "pillow>=10.4.0",
    "google-api-python-client>=2.140.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "numpy>=2.0.1",
    "mypy>=1.11.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/timetracker"]
